backend/blog/serializers.py
from dataclasses import fields
from rest_framework import serializers
from .models import Post

class BlogSerializer(serializers.ModelSerializer):
    class Meta:
        model = Post
        fields = ('title', 'post_text')