backend/requirements.txt
django==3.2
djangorestframework==3.13.1
gunicorn==20.1.*
django-cors-headers
# django-taggit==3.1.0 
# requests==2.28.2
# beautifulsoup4==4.11.2